Attic unused (xyz, masses) -> string utility
