+++ attic/readme.md
@@ -4,6 +4,7 @@ Code, documentation, experiments we want to retain for reference, but that we're
 * `atom_mapping/` -- distance-based atom comparison for rdFMCS
 * `docs/` -- write-up of initial vision for `timemachine`, involving efficient backpropagation through MD trajectories
 * `docking/` -- Docking module that docks uses non-equilibrium switching
+* `io/` -- Input/output utilities
 * `jax_tricks` -- misc. Jax functions
 * `modules` -- misc. modules
    * `reservoir_sampler.py`
