+++ timemachine/fe/utils.py
@@ -12,39 +12,6 @@ def to_md_units(q):
     return q.value_in_unit_system(simtk.unit.md_unit_system)
 
 
-def write(xyz, masses, recenter=True):
-    if recenter:
-        xyz = xyz - np.mean(xyz, axis=0, keepdims=True)
-    buf = str(len(masses)) + "\n"
-    buf += "timemachine\n"
-    for m, (x, y, z) in zip(masses, xyz):
-        if int(round(m)) == 12:
-            symbol = "C"
-        elif int(round(m)) == 14:
-            symbol = "N"
-        elif int(round(m)) == 16:
-            symbol = "O"
-        elif int(round(m)) == 32:
-            symbol = "S"
-        elif int(round(m)) == 35:
-            symbol = "Cl"
-        elif int(round(m)) == 1:
-            symbol = "H"
-        elif int(round(m)) == 31:
-            symbol = "P"
-        elif int(round(m)) == 19:
-            symbol = "F"
-        elif int(round(m)) == 80:
-            symbol = "Br"
-        elif int(round(m)) == 127:
-            symbol = "I"
-        else:
-            raise Exception("Unknown mass:" + str(m))
-
-        buf += symbol + " " + str(round(x, 5)) + " " + str(round(y, 5)) + " " + str(round(z, 5)) + "\n"
-    return buf
-
-
 def convert_uIC50_to_kJ_per_mole(amount_in_uM: float, experiment_temp: float = 298.15) -> float:
     """Convert an IC50 measurement in uM concentrations to kJ/mol.
